Formula	Use 'long' in all integer related places. CCBUG: 127599
svn path=/branches/koffice/1.6/koffice/; revision=628769

+++ kspread/formula.cc
@@ -240,9 +240,9 @@ bool Token::asBoolean() const
   // FIXME check also for i18n version
 }
 
-int Token::asInteger() const
+long Token::asInteger() const
 {
-  if( isInteger() ) return m_text.toInt();
+  if( isInteger() ) return m_text.toLong();
   else return 0;
 }
